Adding Instructions struct
